Adding sudoku solver (extra credit)
It's not much now -- just a bunch of mismatched methods.
I'll try to have it up and running in the next couple of days.

Solver uses backtracking (recursion) to solve the puzzle and keeps
track of possibilities in a 3D array. First 2D are for the puzzle
itself, and other dimension acts as boolean values.

So, sudoku[3][5][0] indicates puzzles's value
All other x (in 1-9) are 0/1 indicating possible/not possible.