Trying out a solution in haskell.
I'm not very confident in deep lore of Haskell what with its
Modads and other things.  But I've got a little bit put here that
may help me in the future for creating a solution. But I'm going to
try to focus my efforts mostly on the C solution. Maybe java if I
can't get it to work...