Complete redesign in redesign.c
I'm totally excited to see that this actually works.  My past
solution (if you can call it that) didn't work at all and had
become too cluttered for me to even look at it.  The problem
was with an improper data structure in a program that heavily
relies on recursion. It was a sad story.

Redesign still has a global sudoku grid, but its not a cube
this time. There's probably a lot of redundant checks that
wouldn't happen if I kept a "possibilities" vector above each
cell on the grid, but oh well, what are you gonna do? Write in
Java? LOL.

Anyway, with modern CPUs this thing is fast as hell. And pretty
clean code too.  I hope you enjoy it, and thanks for reading this
lengthy message.