Haskell solution edits
Now that my C solution is completed, I've turned my attention to
haskell. I have an idea in my head that is very promising and I
hope that it turns out well. It's completely safe and fully
functional. The solve function is loosely lifted from my C code
(at least the idea is). So hopefully it works. We'll see when I
throw a puzzle at it.