Update to solve method for readability
The solve method essentially is the exact same (at least once it
is boiled down to lambda calculus). I've made a significant effort
to grab the code out of the list comprehension and put it into a
series of "where" clauses and not just "let-in" style clauses.

I think the overall outcome makes it much more readable and 1000x
more understandable.  Still though, its a little hard to look at,
but I don't think that will ever change.  But hey, it works!