Edits to solve for readability
Solve basically works the same way, I just hated the way I had
a case/of branch when it wasn't very intuitive. It is now replaced
by a new function which operates on the same principle, which
relies on the if/then/else to change slightly.

No biggies, it looks a lot better and possibly (?) more efficiently.